+++ examples/qt_gui/src/VideoPlayer/VideoRender/VideoPlayerImpl.h
@@ -164,7 +164,7 @@ signals:
     void videoRectChanged(const QRect &rect);
     void playerStateChanged(Stream::PlayerState state);
 
-    void totalTimeRecved(int totalTime);
+    void totalTimeRecved(int64_t totalTime);
     void ptsChanged(double pts);
 
     /*
